// src/views/trade/statistics/type.ts
export interface MyPositionType {
  key: string;
  type: string;
  pnl_usdt: string;
  pnl_usdt_type: string;
  pnl_usdt_percent: string;
  power: number;
  ph: string;
  ph_type: string;
  aprice: string;
  aprice_type: string;
  margin: string;
  margin_type: string;
  risk: string;
  liq_price: string;
  liq_price_type: string;
  tp: string;
  sl: string;
}